backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.1.0
httpx>=0.28.0
pyjwt[crypto]>=2.10.0
supabase>=2.11.0
google-generativeai>=0.8.0
reportlab>=4.2.0
